requirements.txt
feedparser
aiohttp
python-dotenv
requests